+++ app/src/main/java/com/boostcamp/and03/data/datasource/paging/NaverBookSearchPagingSource.kt
@@ -2,7 +2,7 @@ package com.boostcamp.and03.data.datasource.paging
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.boostcamp.and03.data.datasource.remote.NaverBookSearchRemoteDataSource
+import com.boostcamp.and03.data.datasource.remote.search.naver.NaverBookSearchRemoteDataSource
 import com.boostcamp.and03.data.model.response.NaverBookItem
 
 class NaverBookSearchPagingSource(
